Warn if the game directory is empty or invalid
Currently if you don't have it configured (or it's invalid for
wahtever other reason) you get an unhelpful "failed to read Item
EXH" error. Now there is a clearer warning message, which also
prints out the current configuration value.

This is a warning, not an error - in case someone is intentionally
trying to run Kawari without a game path specified.