Fix various issues with display flags
The values were all wrong, which are now fixed. Additionally, the
new Viera ears flag is added which showed up in 7.3. I also made it
so the flag values are actually shown in debug output (which will
help Packet Analyzer especially.)

While testing, I found the ClientSelectData didn't have a spot
for display flags. This was erroneously set to your voice id,
which meant depending on which voice you selected it would randomly
hide parts of your gearset.

See #160