Turn PacketState into ConnectionState, redo how state is handled
I wasn't really happy with PacketState turning into this
super-structure which doesn't really make sense. In the usual case,
there is either a Lobby or Zone connection going on (the two
connections with state) and they have their own data. Another
problem is some other stuff like obsfucation data was tacked on
everywhere and not included in said state.

I redone all of this now, where there's a ConnectionState with
variants, and we now support state-less connections like our
internal server-to-server IPC.