Move zone loading completely to the global server state
This fixes #86, due to the nature of the global server state
instancing all zones it will never load a Zone twice. (At least, in
theory!)

A lot of code has moved from ZoneConnection to Server, but I tested
multiplayer vigorously and it seems to work great. This will also
pave the way for synchronized weather (if changed by a GM) among
other things.

While this doesn't change anything functionally, it would have a
minor positive performance impact as zones are loaded less often
and there's less duplicate copies in memory.