Begin overhauling how event scripts are dispatched
As seen in the previous few commits, we can actually discern a
lot of generic events based on their event ID. Starting small, I
decided to remove our manually curated Aetheryte event list and
replace it with a generic dispatch check.

This also fixes #65, as I also had to figure out a way to access
GameData from the Lua side. We have to use a different Mutex type,
but I think it's worth it.