Begin paving the way for consumable and unlockable items
Previously, items re-used the same action system where each script
was tied to an ID. This worked for something like Fantasia which
is wholly unique in the game, but this model won't work for most
of the unlockables.

It would be simply insane to manually ID every minion item, but
luckily we don't have to. The game already has a way to associate
items with types and now that's used in a new dispatch function. I
already implemented (very loosely) minions and orchestrion items,
but we can extend this into so much more.