Reduce some I/O overhead
Fixes #226, in a way. I wanted to come up with a more generic
caching solution - and we *will* need it at some point but this
should be good enough.

I first updated Physis to make index file reading faster (this is
basically a lookup of paths to their actual data files.)

And then for Kawari itself, I made it so the servers now "preload"
the index files. Since these are a bunch of small files read from
disk, it just makes sense to pay this cost upfront on start-up
instead of on-demand. This *does* mean the servers take marginally
longer to start, but consider this: you would be waiting for that
same blocking I/O read when logging in/traveling to zones anyway.

Finally, I moved even more sheets to the GameData struct so they
aren't read over-and-over. This isn't a perfect solution as it has
to be done manually. You basically want any sheet that is either:
read frequently (Warp) or is so unimaginably large even one read
is quite expensive (Quest)