Attempt to reduce connection timeouts with more graceful exits and shutdowns (#237)
Of course, the messages will still display when the client crashes or loses connection, but it doesn't hurt to make our loops exit gracefully when the client is also gracefully leaving.

I also removed the redundant ToServer::Disconnected message in the ClientZoneIpcData::Disconnected match arm since the end of the zone connection loop sends it anyway. It was causing me some headaches on the party branch, and I didn't notice any obvious regressions with it (I haven't seen zombie players when logging out or force quitting the client).