Redo all of our setup and install documentation
This documentation was kind of a mess, and didn't really have room
to grow. I wanted to add macOS-specific information, but since our
install information is all on one page it would just bloat it.

So I created four separate pages for install instructions: Windows,
macOS and Linux. The fourth page is specifically for local
development. To ensure this doesn't become an unmanageable mess, I
created re-usable Markdown snippets so common steps between all
cases are shared.

While I was at it, I slightly redone the wording for setup in
general so it's now clearer and in the right order. For example we
were telling people to run the servers before they were configured!

Since the hosts file documentation is only relevant for the
Official Launcher, I moved that to the local setup page so people
don't accidentally do that step if it wasn't needed. I also
improved the organization of the local setup page so it isn't a
total information overload and added headers to the
login/registration pages.