Refactor server.rs to use fewer copy-paste loops (#240)
I wrote a handful of helper functions to reduce copy paste, and they allow for either client network to be informed of whatever messages we need to send.

It might be more ideal to perhaps allow passing in a function to call to handle any desired conditional logic in the loop so we could maybe have even less copy paste here, but I couldn't think of a good way to do it at this moment.