Improve server performance again
I was hitting the occasional locked up login again, so its
time to make the server faster! Here's a few improvements that
should compound into a nicer experience:
* Simplified custom IPC connection handling, where it no longer
waits in the background. Since these are always one-shot
connections, there was no point in keeping a loop.
* Cached the Tribe sheet in memory, so its no longer loaded every
time a player logs in.
* Skip loading certain large LGBs that don't seem relevant for
a server.