More improvements towards dungeons
This is kind of a hapzard mess because I was fixing multiple
things at once. But now private/public instances should be
easier to support.

I also found out about the EnterTerritoryEvent packet, which
removes some of our hard-coded code. The remaining zoning
logic is moved to their respective script, which is much
cleaner!