Add a better scheduling system
Previously we (abused) Tokio's tasks and interval system to
delay actions and lose status effects, but this has one obvious
issue: they weren't interruptible!

Now there's a flexible system allowing you to set arbitrary
points in the future, and this is currently used for interruptible
actions and status effects with a duration.