Move GameData to world package, remove from lobby
This is a minor size and start-up time improvement, the lobby
server was the only other thing using GameData and it was only for
the World name.

Since we don't really want to be impersonating retail Worlds,
I removed that feature and let it only be customizable by
modifying Kawari.